+++ app/src/main/java/xterminators/spellingbee/gui/GuiController.java
@@ -9,11 +9,11 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import com.google.gson.JsonSyntaxException;
 
-import xterminators.spellingbee.abstract_ui.Controller;
 import xterminators.spellingbee.model.HelpData;
 import xterminators.spellingbee.model.Puzzle;
 import xterminators.spellingbee.model.PuzzleBuilder;
 import xterminators.spellingbee.model.Rank;
+import xterminators.spellingbee.ui.Controller;
 
 public class GuiController extends Controller {
     /** The view that the user interacts with. */
