+++ app/src/main/java/xterminators/spellingbee/ui/View.java
@@ -1,0 +1,3 @@
+package xterminators.spellingbee.ui;
+
+public abstract class View {}
